refactor structure of workers folder
