+++ source/npm/qsharp/src/language-service/language-service.ts
@@ -20,11 +20,11 @@ import type {
 } from "../../lib/web/qsc_wasm.js";
 import { IProjectHost } from "../main.js";
 import { log } from "../log.js";
-import {
+import type {
   IServiceEventTarget,
   IServiceProxy,
   ServiceProtocol,
-} from "../workers/common.js";
+} from "../workers/types.js";
 type QscWasm = typeof import("../../lib/web/qsc_wasm.js");
 
 export type LanguageServiceDiagnosticEvent = {
